requirements.txt
Pillow==9.3.0
PyPDF2==2.12.1
typing_extensions==4.4.0
